\name{baseModel-package}
\alias{baseModel-package}
\alias{baseModel}
\docType{package}
\title{
\packageTitle{baseModel}
}
\description{
\packageDescription{baseModel}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{baseModel}
\packageIndices{baseModel}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{baseModel}

Maintainer: \packageMaintainer{baseModel}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
